Use ndis_wlan_bssid and ndis_wlan_bssid_ex as per DDK and use whichever one the driver uses (based on length of the structure) when scan items are returned. TI ACX100 driver provides WPA with TKIP, but uses ndis_wlan_bssid structures, which means IEs are not present in the scan items. So the only way to get WPA working with this driver is to use ap_scan=2 with wpa_supplicant so it doesn't scan (and look for IEs in scan data). With this ACX100 works with WPA + TKIP with latest driver for DWL-650+.
Fixed typo (ATTR_SERIALIZZED to ATTR_SERIALIZED).

+++ trunk/ndiswrapper/driver/ntoskernel.h
@@ -1017,15 +1017,13 @@ static inline struct nt_slist *PopEntrySList(nt_slist_header *head,
 #else
 
 #ifndef ll_low
-#define ll_low(x) *(((unsigned int *)&(x)) + 0)
-#define ll_high(x) *(((unsigned int *)&(x)) + 1)
+#define ll_low(x) *(((u32 *)&(x)) + 0)
+#define ll_high(x) *(((u32 *)&(x)) + 1)
 #endif
 
-static inline unsigned long long cmpxchg8b(volatile void *ptr,
-					   unsigned long long old,
-					   unsigned long long new)
+static inline u64 cmpxchg8b(volatile void *ptr, u64 old, u64 new)
 {
-	unsigned long long prev;
+	u64 prev;
 
 	__asm__ __volatile__(LOCK_PREFIX "cmpxchg8b (%0)\n"
 			     : "+r" (ptr),
